Refactored Test Helper classes into their own folder
