+++ backend/task-manager/src/test/java/com/example/task_manager/entity_tests/AdminEntityTest.java
@@ -12,6 +12,7 @@ import org.springframework.test.context.ActiveProfiles;
 
 import com.example.task_manager.entity.Admin;
 import com.example.task_manager.entity.TeamMember;
+import com.example.task_manager.test_helpers.EntityTestHelper;
 
 import jakarta.transaction.Transactional;
