+++ backend/task-manager/src/test/java/com/example/task_manager/entity_tests/AuthInfoEntityTest.java
@@ -9,6 +9,7 @@ import org.springframework.test.context.ActiveProfiles;
 
 import com.example.task_manager.entity.AuthInfo;
 import com.example.task_manager.entity.TeamMember;
+import com.example.task_manager.test_helpers.EntityTestHelper;
 
 import jakarta.persistence.PersistenceException;
 import jakarta.transaction.Transactional;
