+++ backend/task-manager/src/test/java/com/example/task_manager/entity_tests/IsAssignedEntityTest.java
@@ -11,6 +11,7 @@ import com.example.task_manager.entity.IsAssigned;
 import com.example.task_manager.entity.Task;
 import com.example.task_manager.entity.Team;
 import com.example.task_manager.entity.TeamMember;
+import com.example.task_manager.test_helpers.EntityTestHelper;
 
 @DataJpaTest
