+++ backend/task-manager/src/test/java/com/example/task_manager/entity_tests/NotificationEntityTest.java
@@ -12,6 +12,8 @@ import com.example.task_manager.entity.Task;
 import com.example.task_manager.entity.Team;
 import com.example.task_manager.entity.TeamMember;
 import com.example.task_manager.enums.NotificationType;
+import com.example.task_manager.test_helpers.EntityTestHelper;
+
 import jakarta.persistence.PersistenceException;
 
 @DataJpaTest
