+++ backend/task-manager/src/test/java/com/example/task_manager/entity_tests/TeamEntityTest.java
@@ -10,6 +10,7 @@ import com.example.task_manager.entity.IsAssigned;
 import com.example.task_manager.entity.Task;
 import com.example.task_manager.entity.Team;
 import com.example.task_manager.entity.TeamMember;
+import com.example.task_manager.test_helpers.EntityTestHelper;
 import com.example.task_manager.entity.IsMemberOf;
 
 @DataJpaTest
