+++ backend/task-manager/src/test/java/com/example/task_manager/repository_tests/IsMemberOfRepositoryTest.java
@@ -13,6 +13,7 @@ import org.springframework.test.context.ActiveProfiles;
 import com.example.task_manager.entity.IsMemberOf;
 import com.example.task_manager.entity.Team;
 import com.example.task_manager.entity.TeamMember;
+import com.example.task_manager.test_helpers.RepositoryTestHelper;
 
 
 @DataJpaTest
