+++ backend/task-manager/src/test/java/com/example/task_manager/repository_tests/NotificationRepositoryTest.java
@@ -15,6 +15,7 @@ import com.example.task_manager.entity.Task;
 import com.example.task_manager.entity.Team;
 import com.example.task_manager.entity.TeamMember;
 import com.example.task_manager.enums.NotificationType;
+import com.example.task_manager.test_helpers.RepositoryTestHelper;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
