+++ backend/task-manager/src/test/java/com/example/task_manager/repository_tests/TeamRepositoryTest.java
@@ -11,6 +11,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import com.example.task_manager.entity.Team;
 import com.example.task_manager.entity.TeamMember;
+import com.example.task_manager.test_helpers.RepositoryTestHelper;
 
 
 @DataJpaTest
