+++ backend/task-manager/src/test/java/com/example/task_manager/service_tests/AdminServiceTest.java
@@ -24,6 +24,7 @@ import com.example.task_manager.entity.Team;
 import com.example.task_manager.entity.TeamMember;
 import com.example.task_manager.enums.RoleType;
 import com.example.task_manager.enums.TaskPriority;
+import com.example.task_manager.test_helpers.ServiceTestHelper;
 @SpringBootTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @Transactional
