+++ backend/task-manager/src/test/java/com/example/task_manager/service_tests/IsMemberOfServiceTest.java
@@ -11,6 +11,7 @@ import jakarta.transaction.Transactional;
 
 import com.example.task_manager.DTO.TeamDTO;
 import com.example.task_manager.DTO.TeamMemberDTO;
+import com.example.task_manager.test_helpers.ServiceTestHelper;
 
 @SpringBootTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
