+++ backend/task-manager/src/test/java/com/example/task_manager/service_tests/NotificationServiceTest.java
@@ -17,6 +17,7 @@ import com.example.task_manager.entity.Task;
 import com.example.task_manager.entity.Team;
 import com.example.task_manager.entity.TeamMember;
 import com.example.task_manager.enums.NotificationType;
+import com.example.task_manager.test_helpers.ServiceTestHelper;
 
 import jakarta.transaction.Transactional;
